catalog-service/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
ruff==0.1.6

# Optional (for advanced features)
redis==5.0.1
pillow==10.1.0
